pyproject.toml
[project]
name = "desktop-alkozon"
version = "0.1.0"
description = "Desktop app for Alkozon"
requires-python = ">=3.11"
dependencies = [
    "flet>=0.25.0",
    "httpx>=0.27.0",
    "pydantic>=2.10",
    "cryptography>=43.0",
    "python-dotenv>=1.0",
    "keyring>=25.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-mock>=3.10"
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]